requirements.txt

# use `pip3.10`

asyncpraw==7.5.0
discord.py==1.7.3
aiohttp~=3.8.1
prsaw==0.4.0
urllib3~=1.26.2
